+++ src/qt/eventlog.h
@@ -29,7 +29,7 @@ public:
      * @param result Result of the performed call
      * @return success of the operation
      */
-    bool searchTokenTx(int fromBlock, int toBlock, std::string strContractAddress, std::string strSenderAddress, QVariant& result);
+    bool searchTokenTx(int64_t fromBlock, int64_t toBlock, std::string strContractAddress, std::string strSenderAddress, QVariant& result);
 
     /**
      * @brief search Search for log events
@@ -40,12 +40,12 @@ public:
      * @param result Result of the performed call
      * @return success of the operation
      */
-    bool search(int fromBlock, int toBlock, const std::vector<std::string> addresses, const std::vector<std::string> topics, QVariant& result);
+    bool search(int64_t fromBlock, int64_t toBlock, const std::vector<std::string> addresses, const std::vector<std::string> topics, QVariant& result);
 
 private:
     // Set command data
-    void setStartBlock(int fromBlock);
-    void setEndBlock(int toBlock);
+    void setStartBlock(int64_t fromBlock);
+    void setEndBlock(int64_t toBlock);
     void setAddresses(const std::vector<std::string> addresses);
     void setTopics(const std::vector<std::string> topics);
